requirements.txt
Django==3.1.6
djangorestframework==3.12.2
mysqlclient==2.0.3
python-dotenv==0.15.0
celery==5.0.5